requirements.txt
streamlit
langchain
langchain-openai
langgraph
python-dotenv
firebase-admin
pydantic
google-generativeai
requests
